+++ src/crossover/kPoint.js
@@ -1,5 +1,5 @@
 const R = require('ramda');
-const singlePoint = require('./singlePoint');
+const singlePoint = require('./singlePoint')();
 
 const kPoint = k => ([mother, father], random) => {
   const [son, daugher] = R.range(0, k).reduce(