+++ test/unit/crossover/orderOne.test.js
@@ -32,7 +32,7 @@ describe('Order One crossover', () => {
       [8, 2, 1, 3, 4, 5, 6, 7, 9, 0],
     ];
 
-    const result = Genemo.crossover.orderOne([mother, father], random);
+    const result = Genemo.crossover.orderOne()([mother, father], random);
     expect(result).toStrictEqual(offsprings);
   });
 });